src/main/java/cl/transbank/restaurante/controller/LoginController.java
package cl.transbank.restaurante.controller;

import cl.transbank.restaurante.domain.AuthenticationRequest;
import cl.transbank.restaurante.domain.AuthenticationResponse;
import cl.transbank.restaurante.security.AuthenticationService;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RestController;

@RestController
public class LoginController {

    private final AuthenticationService authenticationService;

    public LoginController(AuthenticationService authenticationService) {
        this.authenticationService = authenticationService;
    }

    @PostMapping(path = "/login")
    public ResponseEntity<AuthenticationResponse> login(@RequestBody AuthenticationRequest authRequest) {
        String token = authenticationService.authenticate(authRequest.getUsername(), authRequest.getPassword());
        return ResponseEntity.ok(new AuthenticationResponse(authRequest.getUsername(), token));
    }
}
